Fixed ListView not updating
By using get_queryset, and chaining the three queryset results
(next_early_show, next_late_show and next_workshop) into a list
and returning that, it updates the homepage list.

Also fixed the styleing of the buttons to show the text as white
instead of read;. So you can read it.