Added booking system
Booking system, redirects to login/signup page, allows registered
users to book up to 10 tickets through a simple form. The reservation
is stored in a Model(Reservation), along with the username of the person
who booked it, and the show its realted to.

PROBLEM? I couldn't get the event_date to be connected to the Event_date object
via a ForeignKey, as it wouldn't convert the datetime format properly. Its ok,
but would be better if that was linked.

The system then emails the user (user.email) the confirmation of the reservation
in both HTML and plaintext.