Setup settings and requirements files as per TwoScoops
Almost. Had an issue with the unipath setting to 2 instead of 3
and couldn't get the STATIC_ROOT to play nice with STATICFILES_DIRS
so had to just have STATICFILES_DIRS..is this bad?

Also setup environment variable for SECRET_KEY instead of hardcoding it
and in prod.py setup env variables for DB_USER/PASSWORD/NAME too. Not tested.