Added django_tables2
Installed Django_tables to handle the shoverview page.

Updated requirements/base.txt to include django_tables

Optimised the queries (1054->118!) using select_related() in all
the model queries.

Changed is so you dont have to log in to reserve tickets, but if
you do its slightly easier. Or did I do that yesterday?

Anyway, I changed the view code so that if the first name exists
in the profile it uses that in the booking and if not the user name.

Just noticed the untracked files... will commit them now.